% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_methods.R
\name{edist}
\alias{edist}
\title{Energy Distance Test}
\usage{
edist(x, ix, size, custom)
}
\arguments{
\item{x}{a matrix(n*d) of data points from the first and second group.
each row represent a data point;}

\item{ix}{a permutation of row indices of x.}

\item{size}{a vector of sample sizes;}

\item{custom}{the parameter which is customized. For this test, custom
means a integer used for the number of Lp distance}
}
\value{
Statistics obtained after calculation
\item{T_edist}{The Energy Distance Test statistics}
}
\description{
\code{edist} A Energy Distance Test for equal distributions
}
