% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_methods.R
\name{HT_sqT}
\alias{HT_sqT}
\title{Hotelling’s T-square test}
\usage{
HT_sqT(z, ix, size, custom = NULL)
}
\arguments{
\item{ix}{a permutation of row indices of x.}

\item{size}{a vector of sample sizes;}

\item{custom}{the parameter which is customized. For this test, custom
means nothing}

\item{x}{a matrix(n*d) of data points from the first and second group.
each row represent a data point;}
}
\value{
Statistics obtained after calculation
\item{T_sq}{The EHotelling’s T-square test statistics}
}
\description{
\code{HT_sqT} A Hotelling’s T-square test for equal distributions
}
