% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_methods.R
\name{NNT}
\alias{NNT}
\title{Nearest Neighbor Test}
\usage{
NNT(z, ix, size, custom)
}
\arguments{
\item{z}{a matrix(n*d) of data points from the first and second group.
each row represent a data point;}

\item{ix}{a permutation of row indices of z.}

\item{size}{a vector of sample sizes;}

\item{custom}{the parameter which is customized. For this test, custom
means a integer used for the number of neighbor in nearest neighbor test}
}
\value{
Statistics obtained after calculation
\item{T_nnt}{The nearest neighbor (NN) test statistics}
}
\description{
\code{NNT} A multivariate test for equal distributions is based on
nearest neighbors.The nearest neighbor (NN) tests are a type of test
based on ordered distances between sample elements, which can be
applied when the distributions are continuous.
}
