% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_methods.R
\name{permT}
\alias{permT}
\title{Permutation Test}
\usage{
permT(x, B, ix, size, custom = NULL, Fun)
}
\arguments{
\item{x}{a matrix(n*d) of data points from the first and
second group. each row represent a data point;}

\item{ix}{a permutation of row indices of x.}

\item{size}{a vector of sample sizes;}

\item{custom}{the parameter which is customized. For this test,
custom means parameter used for fun's custom}

\item{Fun}{the function used for this test, including
NNT, edist, HT_sqT, GraphT}

\item{B:}{number of repetitions;}
}
\value{
P-value after calculation
\item{P}{The P-value}
}
\description{
\code{permT} A permutation test for calculate P-value
}
